Out of control
Merovingian: What is the reason? Soon the why and the reason are gone and all that matters is the feeling. This is the nature of the universe. We struggle against it, we fight to deny it; but it is of course a lie. Beneath our poised appearance we are completely out of control.